Add post generation script
This script initializes the generated repository as a git repo.
It furthermore adds an initial commit that includes all the generated
files, with SINTEF as the author via the Team4.0 email.

Finally, it sets up the `origin` remote using the supplied `git_url`
input parameter and changes the branch name to `main`.