Use SINTEF/ci-cd where possible
Update to use SINTEF/ci-cd for the root repo's workflows, where
possible.
Some of the workflows cannot be updated to use the callable workflows
due to them actually working on the template repo or additionally
working on the template repo.
For these, instead of using `invoke`, SINTEF/ci-cd tasks are used.

Some explicit dependencies in the workflows have been moved to a
requirements file under the .github folder.
This requirements file has been added to dependabot.